Move cover controls into separate component
