+++ packages/block-library/src/cover/edit.js
@@ -9,52 +9,23 @@ import namesPlugin from 'colord/plugins/names';
  * WordPress dependencies
  */
 import { useEntityProp, store as coreStore } from '@wordpress/core-data';
+import { useEffect, useRef, useState } from '@wordpress/element';
+import { ResizableBox, Spinner } from '@wordpress/components';
+import { compose } from '@wordpress/compose';
 import {
-	Fragment,
-	useEffect,
-	useRef,
-	useState,
-	useMemo,
-} from '@wordpress/element';
-import {
-	BaseControl,
-	Button,
-	ExternalLink,
-	FocalPointPicker,
-	PanelBody,
-	PanelRow,
-	RangeControl,
-	ResizableBox,
-	Spinner,
-	TextareaControl,
-	ToggleControl,
-	ToolbarButton,
-	__experimentalUseCustomUnits as useCustomUnits,
-	__experimentalToolsPanelItem as ToolsPanelItem,
-	__experimentalUnitControl as UnitControl,
-	__experimentalParseQuantityAndUnitFromRawValue as parseQuantityAndUnitFromRawValue,
-} from '@wordpress/components';
-import { compose, useInstanceId } from '@wordpress/compose';
-import {
-	BlockControls,
 	BlockIcon,
-	InspectorControls,
 	MediaPlaceholder,
-	MediaReplaceFlow,
 	withColors,
 	ColorPalette,
 	useBlockProps,
 	useSetting,
 	useInnerBlocksProps,
 	__experimentalUseGradient,
-	__experimentalPanelColorGradientSettings as PanelColorGradientSettings,
-	__experimentalBlockAlignmentMatrixControl as BlockAlignmentMatrixControl,
-	__experimentalBlockFullHeightAligmentControl as FullHeightAlignmentControl,
 	store as blockEditorStore,
 } from '@wordpress/block-editor';
 import { __ } from '@wordpress/i18n';
 import { useSelect, useDispatch } from '@wordpress/data';
-import { postFeaturedImage, cover as icon } from '@wordpress/icons';
+import { cover as icon } from '@wordpress/icons';
 import { isBlobURL } from '@wordpress/blob';
 import { store as noticesStore } from '@wordpress/notices';
 
@@ -66,13 +37,14 @@ import {
 	attributesFromMedia,
 	IMAGE_BACKGROUND_TYPE,
 	VIDEO_BACKGROUND_TYPE,
-	COVER_MIN_HEIGHT,
 	backgroundImageStyles,
 	dimRatioToClass,
 	isContentPositionCenter,
 	getPositionClassName,
+	mediaPosition,
 } from './shared';
 import useCoverIsDark from './use-cover-is-dark';
+import Controls from './controls';
 
 extend( [ namesPlugin ] );
 
@@ -87,62 +59,6 @@ function getInnerBlocksTemplate( attributes ) {
 			},
 		],
 	];
-}
-
-function CoverHeightInput( {
-	onChange,
-	onUnitChange,
-	unit = 'px',
-	value = '',
-} ) {
-	const instanceId = useInstanceId( UnitControl );
-	const inputId = `block-cover-height-input-${ instanceId }`;
-	const isPx = unit === 'px';
-
-	const units = useCustomUnits( {
-		availableUnits: useSetting( 'spacing.units' ) || [
-			'px',
-			'em',
-			'rem',
-			'vw',
-			'vh',
-		],
-		defaultValues: { px: 430, '%': 20, em: 20, rem: 20, vw: 20, vh: 50 },
-	} );
-
-	const handleOnChange = ( unprocessedValue ) => {
-		const inputValue =
-			unprocessedValue !== ''
-				? parseFloat( unprocessedValue )
-				: undefined;
-
-		if ( isNaN( inputValue ) && inputValue !== undefined ) {
-			return;
-		}
-		onChange( inputValue );
-	};
-
-	const computedValue = useMemo( () => {
-		const [ parsedQuantity ] = parseQuantityAndUnitFromRawValue( value );
-		return [ parsedQuantity, unit ].join( '' );
-	}, [ unit, value ] );
-
-	const min = isPx ? COVER_MIN_HEIGHT : 0;
-
-	return (
-		<BaseControl label={ __( 'Minimum height of cover' ) } id={ inputId }>
-			<UnitControl
-				id={ inputId }
-				isResetValueOnUnitChange
-				min={ min }
-				onChange={ handleOnChange }
-				onUnitChange={ onUnitChange }
-				style={ { maxWidth: 80 } }
-				units={ units }
-				value={ computedValue }
-			/>
-		</BaseControl>
-	);
 }
 
 const RESIZABLE_BOX_ENABLE_OPTION = {
@@ -188,10 +104,6 @@ function ResizableCover( {
 			{ ...props }
 		/>
 	);
-}
-
-function mediaPosition( { x, y } ) {
-	return `${ Math.round( x * 100 ) }% ${ Math.round( y * 100 ) }%`;
 }
 
 /**
@@ -287,66 +199,9 @@ function CoverEdit( {
 		blockEditorStore
 	);
 	const { createErrorNotice } = useDispatch( noticesStore );
-	const {
-		gradientClass,
-		gradientValue,
-		setGradient,
-	} = __experimentalUseGradient();
+	const { gradientClass, gradientValue } = __experimentalUseGradient();
 	const onSelectMedia = attributesFromMedia( setAttributes, dimRatio );
 	const isUploadingMedia = isTemporaryMedia( id, url );
-
-	const [ prevMinHeightValue, setPrevMinHeightValue ] = useState( minHeight );
-	const [ prevMinHeightUnit, setPrevMinHeightUnit ] = useState(
-		minHeightUnit
-	);
-	const isMinFullHeight = minHeightUnit === 'vh' && minHeight === 100;
-
-	const toggleMinFullHeight = () => {
-		if ( isMinFullHeight ) {
-			// If there aren't previous values, take the default ones.
-			if ( prevMinHeightUnit === 'vh' && prevMinHeightValue === 100 ) {
-				return setAttributes( {
-					minHeight: undefined,
-					minHeightUnit: undefined,
-				} );
-			}
-
-			// Set the previous values of height.
-			return setAttributes( {
-				minHeight: prevMinHeightValue,
-				minHeightUnit: prevMinHeightUnit,
-			} );
-		}
-
-		setPrevMinHeightValue( minHeight );
-		setPrevMinHeightUnit( minHeightUnit );
-
-		// Set full height.
-		return setAttributes( {
-			minHeight: 100,
-			minHeightUnit: 'vh',
-		} );
-	};
-
-	const toggleParallax = () => {
-		setAttributes( {
-			hasParallax: ! hasParallax,
-			...( ! hasParallax ? { focalPoint: undefined } : {} ),
-		} );
-	};
-
-	const toggleIsRepeated = () => {
-		setAttributes( {
-			isRepeated: ! isRepeated,
-		} );
-	};
-
-	const toggleUseFeaturedImage = () => {
-		setAttributes( {
-			useFeaturedImage: ! useFeaturedImage,
-			dimRatio: dimRatio === 100 ? 50 : dimRatio,
-		} );
-	};
 
 	const onUploadError = ( message ) => {
 		createErrorNotice( Array.isArray( message ) ? message[ 2 ] : message, {
@@ -394,203 +249,12 @@ function CoverEdit( {
 	};
 
 	const hasBackground = !! ( url || overlayColor.color || gradientValue );
-	const showFocalPointPicker =
-		isVideoBackground ||
-		( isImageBackground && ( ! hasParallax || isRepeated ) );
-
-	const imperativeFocalPointPreview = ( value ) => {
-		const [ styleOfRef, property ] = isDarkElement.current
-			? [ isDarkElement.current.style, 'objectPosition' ]
-			: [ ref.current.style, 'backgroundPosition' ];
-		styleOfRef[ property ] = mediaPosition( value );
-	};
 
 	const hasInnerBlocks = useSelect(
 		( select ) =>
 			select( blockEditorStore ).getBlock( clientId ).innerBlocks.length >
 			0,
 		[ clientId ]
-	);
-
-	const controls = (
-		<>
-			<BlockControls group="block">
-				<BlockAlignmentMatrixControl
-					label={ __( 'Change content position' ) }
-					value={ contentPosition }
-					onChange={ ( nextPosition ) =>
-						setAttributes( {
-							contentPosition: nextPosition,
-						} )
-					}
-					isDisabled={ ! hasInnerBlocks }
-				/>
-				<FullHeightAlignmentControl
-					isActive={ isMinFullHeight }
-					onToggle={ toggleMinFullHeight }
-					isDisabled={ ! hasInnerBlocks }
-				/>
-			</BlockControls>
-			<BlockControls group="other">
-				<ToolbarButton
-					icon={ postFeaturedImage }
-					label={ __( 'Use featured image' ) }
-					isPressed={ useFeaturedImage }
-					onClick={ toggleUseFeaturedImage }
-				/>
-				{ ! useFeaturedImage && (
-					<MediaReplaceFlow
-						mediaId={ id }
-						mediaURL={ url }
-						allowedTypes={ ALLOWED_MEDIA_TYPES }
-						accept="image/*,video/*"
-						onSelect={ onSelectMedia }
-						name={ ! url ? __( 'Add Media' ) : __( 'Replace' ) }
-					/>
-				) }
-			</BlockControls>
-			<InspectorControls>
-				{ !! url && (
-					<PanelBody title={ __( 'Media settings' ) }>
-						{ isImageBackground && (
-							<Fragment>
-								<ToggleControl
-									label={ __( 'Fixed background' ) }
-									checked={ hasParallax }
-									onChange={ toggleParallax }
-								/>
-
-								<ToggleControl
-									label={ __( 'Repeated background' ) }
-									checked={ isRepeated }
-									onChange={ toggleIsRepeated }
-								/>
-							</Fragment>
-						) }
-						{ showFocalPointPicker && (
-							<FocalPointPicker
-								label={ __( 'Focal point picker' ) }
-								url={ url }
-								value={ focalPoint }
-								onDragStart={ imperativeFocalPointPreview }
-								onDrag={ imperativeFocalPointPreview }
-								onChange={ ( newFocalPoint ) =>
-									setAttributes( {
-										focalPoint: newFocalPoint,
-									} )
-								}
-							/>
-						) }
-						{ ! useFeaturedImage &&
-							url &&
-							isImageBackground &&
-							isImgElement && (
-								<TextareaControl
-									label={ __(
-										'Alt text (alternative text)'
-									) }
-									value={ alt }
-									onChange={ ( newAlt ) =>
-										setAttributes( { alt: newAlt } )
-									}
-									help={
-										<>
-											<ExternalLink href="https://www.w3.org/WAI/tutorials/images/decision-tree">
-												{ __(
-													'Describe the purpose of the image'
-												) }
-											</ExternalLink>
-											{ __(
-												'Leave empty if the image is purely decorative.'
-											) }
-										</>
-									}
-								/>
-							) }
-						<PanelRow>
-							<Button
-								variant="secondary"
-								isSmall
-								className="block-library-cover__reset-button"
-								onClick={ () =>
-									setAttributes( {
-										url: undefined,
-										id: undefined,
-										backgroundType: undefined,
-										focalPoint: undefined,
-										hasParallax: undefined,
-										isRepeated: undefined,
-										useFeaturedImage: false,
-									} )
-								}
-							>
-								{ __( 'Clear Media' ) }
-							</Button>
-						</PanelRow>
-					</PanelBody>
-				) }
-				<PanelColorGradientSettings
-					__experimentalHasMultipleOrigins
-					__experimentalIsRenderedInSidebar
-					title={ __( 'Overlay' ) }
-					initialOpen={ true }
-					settings={ [
-						{
-							colorValue: overlayColor.color,
-							gradientValue,
-							onColorChange: setOverlayColor,
-							onGradientChange: setGradient,
-							label: __( 'Color' ),
-						},
-					] }
-				>
-					<RangeControl
-						label={ __( 'Opacity' ) }
-						value={ dimRatio }
-						onChange={ ( newDimRation ) =>
-							setAttributes( {
-								dimRatio: newDimRation,
-							} )
-						}
-						min={ 0 }
-						max={ 100 }
-						step={ 10 }
-						required
-					/>
-				</PanelColorGradientSettings>
-			</InspectorControls>
-			<InspectorControls __experimentalGroup="dimensions">
-				<ToolsPanelItem
-					hasValue={ () => !! minHeight }
-					label={ __( 'Minimum height' ) }
-					onDeselect={ () =>
-						setAttributes( {
-							minHeight: undefined,
-							minHeightUnit: undefined,
-						} )
-					}
-					resetAllFilter={ () => ( {
-						minHeight: undefined,
-						minHeightUnit: undefined,
-					} ) }
-					isShownByDefault={ true }
-					panelId={ clientId }
-				>
-					<CoverHeightInput
-						value={ minHeight }
-						unit={ minHeightUnit }
-						onChange={ ( newMinHeight ) =>
-							setAttributes( { minHeight: newMinHeight } )
-						}
-						onUnitChange={ ( nextUnit ) =>
-							setAttributes( {
-								minHeightUnit: nextUnit,
-							} )
-						}
-					/>
-				</ToolsPanelItem>
-			</InspectorControls>
-		</>
 	);
 
 	const ref = useRef();
@@ -614,10 +278,28 @@ function CoverEdit( {
 		}
 	);
 
+	const currentSettings = {
+		isVideoBackground,
+		isImageBackground,
+		isDarkElement,
+		hasInnerBlocks,
+		url,
+		isImgElement,
+		overlayColor,
+	};
+
 	if ( ! hasInnerBlocks && ! hasBackground ) {
 		return (
 			<>
-				{ controls }
+				<Controls
+					attributes={ attributes }
+					setAttributes={ setAttributes }
+					clientId={ clientId }
+					setOverlayColor={ setOverlayColor }
+					ref={ ref }
+					onSelectMedia={ onSelectMedia }
+					currentSettings={ currentSettings }
+				/>
 				<div
 					{ ...blockProps }
 					className={ classnames(
@@ -677,7 +359,15 @@ function CoverEdit( {
 
 	return (
 		<>
-			{ controls }
+			<Controls
+				attributes={ attributes }
+				setAttributes={ setAttributes }
+				clientId={ clientId }
+				setOverlayColor={ setOverlayColor }
+				ref={ ref }
+				onSelectMedia={ onSelectMedia }
+				currentSettings={ currentSettings }
+			/>
 			<div
 				{ ...blockProps }
 				className={ classnames( classes, blockProps.className ) }
